provide client method to setup all queues and queue policies
Setting up queue policies on demand in publisher and subscriber is OK
for a small number of queues, publishers and subscribers. But the HTTP
API of RabbitMQ doesn't scale all that well, so that a large number
of HTPP calls to set up queue policies can in fact crash a server.