remove queue bindings which are no longer part of the queue declaration
ATTENTION: this assumes that queue declarations are not shared between
different capplications.

Note that the implementation is only eventually consistent in that
during an application deployment it is possible that old clients can cause
deleted bindings to reappear due to queue declarations. As a consequence,
consumers need to be able to handle messages messages routed previously,
until all clients have been restarted.