delayed expiry of status keys in Redis
By actually tracking handler execution in one of our applications,
we have proof that every once in a while business logic (i.e. the
process method of a message handler) gets executed more than once.

This patch is an experiment to see whether we can eliminate these
duplicates by shifting the responsibility of deleting the corresponding
status key to Redis itself (with a configurable grace period).

One possible source of duplicate handler executions is the application
crashing while the ack message for the successfully processed RabbitMQ
message is still sitting in a TCP send buffer. This complication would
be handled by the patch by tuning the grace period (in many cases).
However, to fix it in all cases we'd need to keep the status keys which
act as tombstones, for as long as a message hasn't officially expired.