PROBLEM: handler for simple messages can be executed twice if acks are lost
SOLUTION: similar to redundant messages, we set a completion status in
Redis after a handler has been executed. The value is always the
string "completed", regardless of whether the Handler has crashed or
not. The status key's expiry time is set to the globally configured
redis_status_key_expiry_interval. We make sure to set both value and
expiry with a single Redis command to avoid potential memory leaks in
Redis.