avoid put call when updating queue policies
It seems the additional call to get the current definition of a policy
is to be preferred over relying on the idempotency of the PUT call.
This helps when adding a new fresh server to a beetle cluster: import
the definitions of one of the existing nodes on the fresh node before
actually adding it to the list of servers in the client's beetle config.