Thread local channels (#116)
With this change we make sure that publishers don't share channels across threads.
In our applications, we usually access the publisher through a global object, and it's not clear in which thread of execution it will be used. Normally we would just advise to change the approach and i.e. use one Client per thread, but unfortunately, we don't control how the code is currently used. This change makes sure, that the rare cases where threads might be used and the client is shared, don't run into out of order frame problems.

---------

Co-authored-by: Daniel Goncharov <danil.goncharov@gmail.com>
Co-authored-by: Bruno Escherl <bruno.escherl@xing.com>