Adapt policy updates (#130)
* Allow to overwrite updates

* Give full control over queue policy updates

This updates the code to give full control over the policy update
request to the caller. Also if available we updated the queue policy
options payload to contain information about the queue type.

* Add method to retrieve current queue properties on server

This allows us to deduce the correct `apply-to` in the policy update manager.

* Don't send queue type anymore

We can always ask the server

* remove relic from older change