Allow to disable lazy queue setup (#141)
This allows to disable the binding of queues for exchanges,
on first publish to an exchange.

The declaration of those queues can be costly as it involves
communication with the broker.

If you want to disable this you can optimize the effect, but
on the flip-side you have to make sure that queues are bound
on the server before.