Actually disable autorecovery on network failure (#145)
This actually disables bunny auto-recovery, since it is fundamentally incompatible with the failure model of beetle publisher.
It disables auto-recovery for network errors.
It disables auto-recovery for forced closes from the server.
It disables the recovery sleep interval which is used currently in a buggy bunny version even when auto recovery is disabled.