Publisher handles all errors in publish the same
In all cases we want to play safe, reset the publishing state, and retry if possible.
Before this change there was the chance, that unknown errors that could
occur would not be handled correctly, again leaving room rogue threads.

This simplifies the error handling of publish to crash-recovery for all
errors.