Make sure we actually timeout on bunny connect via TLS (#154)
This adds additional timeouts for TLS connections, so that when servers are not available we don't wait more than the timeouts configured for the publisher.

This is done by also introducing a ruby timeout on top of the socket timeout which will make sure, that long TLS negotiations after TCP connect are interrupted when the timeout has elapsed.

We make sure to cleanup any partial state of the connection in that case.